fix: (#56) 로그아웃시 유저아이디가 있는지 확인

+++ src/main/java/com/example/semiwiki_backend/domain/auth/service/LogoutService.java
@@ -1,5 +1,7 @@
 package com.example.semiwiki_backend.domain.auth.service;
 
+import com.example.semiwiki_backend.domain.auth.exception.AccountNotFoundException;
+import com.example.semiwiki_backend.domain.user.repository.UserRepository;
 import com.example.semiwiki_backend.global.security.jwt.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -9,9 +11,16 @@ import org.springframework.stereotype.Service;
 public class LogoutService {
 
     private final JwtTokenProvider jwtTokenProvider;
+    private final UserRepository userRepository;
 
     public void execute(String accountId){
-        jwtTokenProvider.deleteRefreshToken(accountId);
+        if(userRepository.existsByAccountId(accountId)){
+            jwtTokenProvider.deleteRefreshToken(accountId);
+        }
+        else {
+            throw new AccountNotFoundException();
+        }
+
 
     }
 }
